Created CPT class
Moved methods from Node to the CPT class. After the changes I reran the tests for Node. All of the tests passed.
