+++ bayesian_network/bayesian_network/graph/node.py
@@ -83,32 +83,6 @@ class Node:
                 return parent
         return None
 
-    def get_all_combination(self, l):
-        """
-        get_all_combination takes in a list of items, then returns all possible combinations of each item.
-        :param l: list of items
-        :return: list of combinations
-        """
-        combination_list = []
-        for i in list(itertools.product(l, repeat=len(self._parents))):
-            combination_list.append(i)
-
-        return combination_list
-
-    def get_combination_probability(self, filtered_data, predict):
-        """
-        get_combination_probability will return the probability of the node's state given it's parents' states.
-        :param filtered_data: DataFrame
-        :param predict: state of Node
-        :return: float
-        """
-        number_of_rows = filtered_data.shape[0]
-        if number_of_rows == 0:
-            return 0
-        occurrence = len(filtered_data[filtered_data[self._name] == predict])
-
-        return occurrence / number_of_rows
-
     def add_child(self, child):
         """
         add_child will add a child to the end of _children. If _children is None then a new list will be created with
@@ -136,60 +110,4 @@ class Node:
             if child.get_name() == name_of_child:
                 self._children.remove(child)
 
-    def update_cp_table(self, data, scale):
-        """
-        update_probability_table will up a DataFrame that will represent the Node's Conditional Probability Table
-        (CP Table). This DataFrame will contain all of the probability of the Node's states given the parent's states.
-        :param data: DataFrame
-        :param scale: List of strings
-        """
-        cpt = self.create_data_frame_columns(scale)
-        # When a Node has no parents then filtering based on parent states is not needed.
-        if len(self._parents) == 0:
-            row = []
-            for i in range(0, len(scale)):
-                row.append(self.get_combination_probability(data, scale[i]))
-            cpt.loc[0] = row
 
-            self._cp_table = cpt
-        else:
-            combo = self.get_all_combination(scale)
-
-            for i in range(0, len(combo)):
-                row = []
-                row.extend(combo[i])
-                for value in scale:
-                    row.append(self.get_combination_probability(self.filter_data(data, combo[i]), value))
-                cpt.loc[i] = row
-
-            self._cp_table = cpt
-
-    def create_data_frame_columns(self, scale):
-        """
-        create_data_frame_columns will add all the node's parent's names as columns in the DataFrame. After the parent's
-        each value in the scale will also be a column after the parents name.
-        :param scale: list of Strings
-        :return: DataFrame
-        """
-        columns = []
-
-        if len(self._parents) != 0:
-            for parent in self._parents:
-                columns.append(parent.get_name())
-
-        columns.extend(scale)
-
-        return pd.DataFrame(columns=columns)
-
-    def filter_data(self, data, combination):
-        """
-        filter_data will filter the data frame based on the combination of the parents states.
-        :param data: Data Frame. Pre condition: Node must have at least one parent
-        :param combination: list of strings
-        :return: Data Frame
-        """
-        df = data
-
-        for i in range(0, len(self._parents)):
-            df = df[df[self._parents[i].get_name()] == combination[i]]
-        return df
